+++ common/src/main/java/concerrox/emixx/util/SearchWorkerBridge.java
@@ -1,0 +1,5 @@
+package concerrox.emixx.util;
+
+public interface SearchWorkerBridge {
+    String emixx$getQuery();
+}