+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.7] - 2026-02-10
+
+### Added
+
+- Stack group names now show up in EMI search results.
+
+### Fixed
+
+- Fixed highlights persisting in the EMI menu.
+- Fixed stackgroups not recursively populating from item tags.
+
 ## [1.2.6] - 2026-02-09
 
 ### Fixed